refactor(sip)+feat: extract SipReferHandler + CF-045 Slice 2 pending-state (RFC 6665)
Two changes on the CF-045 line:

1. Extract Class: REFER handling (HandleReferAsync + the sipfrag NOTIFY sender)
   moved verbatim out of SipCallSessionInboundService (was 999 lines, at the 1000
   ceiling) into a new SipReferHandler. The inbound service holds one _refer and
   delegates the REFER dispatch; behaviour-preserving, no partial class. Inbound
   service is now 883 lines, creating headroom for further REFER work.

2. CF-045 Slice 2 — RFC 6665 pending-state: new IReferSubscription.ReportPending()
   makes the immediate NOTIFY carry Subscription-State: pending instead of active.
   Effective only when signalled synchronously inside the transfer handler (before
   the first NOTIFY); the first subsequent progress report transitions to active;
   no-op once started/terminated, and active→pending is impossible. Default (no
   ReportPending) is unchanged from Slice 1 (active/100).

Tests: +2 (pending→active→terminated; pending-only single NOTIFY) → 8 total.
Revert-verify: forcing the initial state to active turned both pending tests RED.
Release 0 warn; Arch 12/12 (inbound 883, handler 142, subscription 158); Core
1478/1478; Client 86/86. Review APPROVED.

Deferred follow-up: SDK-side auto-timeout (terminated;reason=timeout) for an
accepted REFER whose consumer never reports — needs a session-lifetime-bound timer.
